refactor(telegram): extract s4 gating/media clusters into mixin modules (epic #78791)
