+++ plugins/platforms/telegram/adapter.py
@@ -38,23 +38,6 @@ def _redact_telegram_error_text(error: object) -> str:
         return "<telegram error redacted>"
 
 
-def _scoped_gate_env(name: str, default: str = "") -> str:
-    """Read a TELEGRAM_*/GATEWAY_* authorization gate env var per-profile.
-
-    Under gateway.multiplex_profiles the process env is first-writer-wins
-    (the YAML→env bridge in ``_apply_yaml_config``), so a raw ``os.getenv``
-    can return ANOTHER profile's allowlist (issue #72348, Telegram mirror).
-    Reads the active profile's secret scope when installed; falls back to
-    ``os.getenv`` outside multiplex — identical single-profile behavior.
-    """
-    try:
-        from gateway.authz_mixin import _platform_gate_env
-
-        return _platform_gate_env(name, default)
-    except Exception:
-        return (os.getenv(name) or default).strip()
-
-
 def _consume_abandoned_task(task: asyncio.Task) -> None:
     """Observe a detached task's terminal exception to avoid noisy loop logs."""
     try:
@@ -323,74 +306,6 @@ _TELEGRAM_IMAGE_EXT_TO_MIME = {
     ".gif": "image/gif",
 }
 
-def _coerce_duration_seconds(value: Any) -> Optional[int]:
-    """Round a raw length to whole positive seconds, or None if unusable."""
-    try:
-        secs = int(round(float(value)))
-    except (TypeError, ValueError):
-        return None
-    return secs if secs > 0 else None
-
-
-def _probe_voice_duration_seconds(path: str) -> Optional[int]:
-    """Best-effort audio length in whole seconds for outgoing voice/audio.
-
-    Telegram only auto-derives a clip's duration from container metadata for
-    short recordings; longer ones (roughly 5 min+) are sent with duration 0
-    and render as ``0:00`` in the player. We read the length locally and pass
-    it explicitly so the bubble shows the real time.
-
-    Mirrors ``gateway.run._probe_audio_duration``: stdlib ``wave`` for WAV,
-    then mutagen for OGG/Opus/MP3/M4A metadata, then an ``ffprobe`` fallback.
-    All three are optional — when none can read the file we return ``None``
-    and the caller omits ``duration``, falling back to Telegram's own
-    (possibly absent) metadata, i.e. the prior behavior. Blocking (mutagen
-    read + ffprobe subprocess), so call it via ``asyncio.to_thread``.
-    """
-    ext = os.path.splitext(path)[1].lower()
-
-    if ext == ".wav":
-        try:
-            import wave
-
-            with wave.open(path, "rb") as wf:
-                rate = wf.getframerate() or 0
-                if rate:
-                    secs = _coerce_duration_seconds(wf.getnframes() / float(rate))
-                    if secs is not None:
-                        return secs
-        except Exception:
-            pass
-
-    try:
-        import mutagen
-
-        audio = mutagen.File(path)
-        secs = _coerce_duration_seconds(
-            getattr(getattr(audio, "info", None), "length", None)
-        )
-        if secs is not None:
-            return secs
-    except Exception:
-        pass
-
-    try:
-        import shutil
-        import subprocess
-
-        if shutil.which("ffprobe"):
-            proc = subprocess.run(
-                ["ffprobe", "-v", "error", "-show_entries", "format=duration",
-                 "-of", "default=noprint_wrappers=1:nokey=1", path],
-                capture_output=True, text=True, encoding="utf-8", errors="replace", timeout=5,
-            )
-            if proc.returncode == 0:
-                return _coerce_duration_seconds(proc.stdout.strip())
-    except Exception:
-        pass
-
-    return None
-
 
 def check_telegram_requirements() -> bool:
     """Check if Telegram dependencies are available.
@@ -447,16 +362,6 @@ def check_telegram_requirements() -> bool:
     return True
 
 
-# Matches every character that MarkdownV2 requires to be backslash-escaped
-# when it appears outside a code span or fenced code block.
-_MDV2_ESCAPE_RE = re.compile(r'([_*\[\]()~`>#\+\-=|{}.!\\])')
-
-
-def _escape_mdv2(text: str) -> str:
-    """Escape Telegram MarkdownV2 special characters with a preceding backslash."""
-    return _MDV2_ESCAPE_RE.sub(r'\\\1', text)
-
-
 def _strip_mdv2(text: str) -> str:
     """Strip MarkdownV2 escape backslashes to produce clean plain text.
 
@@ -507,8 +412,6 @@ def _separate_chunk_indicator_from_fence(text: str) -> str:
 
 from gateway.platforms.helpers import (
     TABLE_SEPARATOR_RE as _TABLE_SEPARATOR_RE,
-    compile_mention_patterns,
-    convert_table_to_bullets as _wrap_markdown_tables,
 )
 
 
@@ -598,13 +501,7 @@ _POLLING_ERROR_TASK_STUCK_TIMEOUT = 300.0
 # A generation is not healthy until the dedicated getUpdates request returns
 # successfully. This exceeds a normal long-poll cycle for healthy idle bots.
 _POLLING_PROGRESS_TIMEOUT = 60.0
-# Telegram transcodes an uploaded video before it answers sendVideo, so the
-# wait for the response is unrelated to how fast the bytes went out and can
-# outlast the 20s read timeout the rest of the Bot API is tuned for. Only
-# media sends take this longer budget; ordinary calls keep the short one so a
-# dead request is still noticed quickly. Kept modest deliberately — this is
-# also how long a user waits to be told the attachment failed.
-_MEDIA_SEND_READ_TIMEOUT = 60.0
+
 _POLLING_GENERATION_CONTEXT: ContextVar[Optional[int]] = ContextVar(
     "telegram_polling_generation", default=None
 )
@@ -614,7 +511,34 @@ class _PollingLifecycleAbort(RuntimeError):
     """Internal control flow for polling startup fenced by teardown."""
 
 
-class TelegramAdapter(BasePlatformAdapter):
+from plugins.platforms.telegram.bot_identity_mixin import BotIdentityMixin
+from plugins.platforms.telegram.callbacks_mixin import CallbacksMixin
+from plugins.platforms.telegram.formatting_mixin import (
+    FormattingMixin,
+    _escape_mdv2,
+    _wrap_markdown_tables,
+)
+from plugins.platforms.telegram.group_gating_mixin import (
+    GroupGatingMixin,
+    _scoped_gate_env,
+)
+from plugins.platforms.telegram.media_send_mixin import (
+    MediaSendMixin,
+    _MEDIA_SEND_READ_TIMEOUT,
+    _coerce_duration_seconds,
+    _probe_voice_duration_seconds,
+)
+from plugins.platforms.telegram.typing_mixin import TypingMixin
+
+class TelegramAdapter(
+    GroupGatingMixin,
+    MediaSendMixin,
+    TypingMixin,
+    FormattingMixin,
+    BotIdentityMixin,
+    CallbacksMixin,
+    BasePlatformAdapter,
+):
     """
     Telegram bot adapter.
 
@@ -6282,411 +6206,6 @@ class TelegramAdapter(BasePlatformAdapter):
             # Catch-all (e.g. page counter button "mx:noop")
             await query.answer()
 
-    async def _notify_clarify_expired(self, query, user_display: str) -> None:
-        """Tell the user a clarify tap arrived too late to be delivered.
-
-        Fires when the clarify entry was evicted by ``clarify_timeout`` or the
-        gateway restarted between asking and the tap. In both cases the agent
-        thread is no longer waiting, so the tap would otherwise leave a
-        misleading ✓ (or an "awaiting typed response" prompt) on a button the
-        agent never receives.
-        """
-        try:
-            await query.answer(text="⚠️ This prompt expired — please /retry.")
-        except Exception:
-            pass
-        try:
-            await query.edit_message_text(
-                text=(
-                    f"❓ {_html.escape(query.message.text or '')}\n\n"
-                    "<i>⚠️ This question expired or the session reset — please /retry.</i>"
-                ),
-                parse_mode=ParseMode.HTML,
-                reply_markup=None,
-            )
-        except Exception:
-            pass
-
-    async def _handle_callback_query(
-        self, update: "Update", context: "ContextTypes.DEFAULT_TYPE"
-    ) -> None:
-        """Handle inline keyboard button clicks."""
-        query = update.callback_query
-        if not query or not query.data:
-            return
-        data = query.data
-        query_message = getattr(query, "message", None)
-        query_chat_id = getattr(query_message, "chat_id", None)
-        query_chat = getattr(query_message, "chat", None)
-        query_chat_type = getattr(query_chat, "type", None)
-        query_thread_id = getattr(query_message, "message_thread_id", None)
-        query_user_name = getattr(query.from_user, "first_name", None)
-
-        # --- Model picker callbacks ---
-        if data.startswith(("mp:", "mpg:", "mpv:", "mm:", "mc:", "mb", "mx", "mg:")):
-            chat_id = str(query.message.chat_id) if query.message else None
-            if chat_id:
-                await self._handle_model_picker_callback(query, data, chat_id)
-            return
-
-        # --- Generic choice picker callbacks (/reasoning, /fast) ---
-        if data.startswith("cp:"):
-            chat_id = str(query.message.chat_id) if query.message else None
-            if chat_id:
-                await self._handle_choice_picker_callback(query, data, chat_id)
-            return
-
-        # --- Gmail-triage callbacks (gt:verb:arg) ---
-        if data.startswith("gt:"):
-            await self._handle_gmail_triage_callback(
-                query,
-                data,
-                query_chat_id=query_chat_id,
-                query_chat_type=query_chat_type,
-                query_thread_id=query_thread_id,
-                query_user_name=query_user_name,
-            )
-            return
-
-        # --- Exec approval callbacks (ea:choice:id) ---
-        if data.startswith("ea:"):
-            parts = data.split(":", 2)
-            if len(parts) == 3:
-                choice = parts[1]  # once, session, always, deny
-                try:
-                    approval_id = int(parts[2])
-                except (ValueError, IndexError):
-                    await query.answer(text="Invalid approval data.")
-                    return
-
-                # Only authorized users may click approval buttons.
-                caller_id = str(getattr(query.from_user, "id", ""))
-                if not self._is_callback_user_authorized(
-                    caller_id,
-                    chat_id=query_chat_id,
-                    chat_type=str(query_chat_type) if query_chat_type is not None else None,
-                    thread_id=str(query_thread_id) if query_thread_id is not None else None,
-                    user_name=query_user_name,
-                ):
-                    await query.answer(text="⛔ You are not authorized to approve commands.")
-                    return
-
-                session_key = self._approval_state.pop(approval_id, None)
-                if not session_key:
-                    await query.answer(text="This approval has already been resolved.")
-                    return
-
-                user_display = getattr(query.from_user, "first_name", "User")
-
-                # Resolve the approval FIRST — unblocks the agent thread.
-                # Rendering happens after so the message reflects what
-                # actually occurred: a tap that lands after the approval
-                # wait timed out (count == 0) must NOT claim "Approved" —
-                # the command was already denied and will not run (#63501
-                # regression follow-up: 60s waits made stale taps common).
-                try:
-                    from tools.approval import resolve_gateway_approval
-                    count = resolve_gateway_approval(session_key, choice)
-                    logger.info(
-                        "Telegram button resolved %d approval(s) for session %s (choice=%s, user=%s)",
-                        count, session_key, choice, user_display,
-                    )
-                except Exception as exc:
-                    logger.error("Failed to resolve gateway approval from Telegram button: %s", exc)
-                    count = 0
-
-                if count:
-                    # Map choice to human-readable label
-                    label_map = {
-                        "once": "✅ Approved once",
-                        "session": "✅ Approved for session",
-                        "always": "✅ Approved permanently",
-                        "deny": "❌ Denied",
-                    }
-                    label = label_map.get(choice, "Resolved")
-                    edit_text = f"{label} by {user_display}"
-                else:
-                    label = "⌛ Approval expired"
-                    edit_text = (
-                        f"{label} — no command was waiting. "
-                        f"It already timed out (and was denied) or was resolved elsewhere."
-                    )
-
-                await query.answer(text=label)
-
-                # Edit message to show decision, remove buttons
-                try:
-                    await query.edit_message_text(
-                        text=self.format_message(edit_text),
-                        parse_mode=ParseMode.MARKDOWN_V2,
-                        reply_markup=None,
-                    )
-                except Exception:
-                    pass  # non-fatal if edit fails
-
-                # Resume the typing indicator — paused when the approval was
-                # sent (gateway/run.py).  The text /approve and /deny paths
-                # call resume_typing_for_chat here too; without it, typing
-                # stays paused for the rest of the turn after an inline
-                # button click.
-                if count and query_chat_id is not None:
-                    self.resume_typing_for_chat(str(query_chat_id))
-            return
-
-        # --- Slash-confirm callbacks (sc:choice:confirm_id) ---
-        if data.startswith("sc:"):
-            parts = data.split(":", 2)
-            if len(parts) == 3:
-                choice = parts[1]  # once, always, cancel
-                confirm_id = parts[2]
-
-                caller_id = str(getattr(query.from_user, "id", ""))
-                if not self._is_callback_user_authorized(
-                    caller_id,
-                    chat_id=query_chat_id,
-                    chat_type=str(query_chat_type) if query_chat_type is not None else None,
-                    thread_id=str(query_thread_id) if query_thread_id is not None else None,
-                    user_name=query_user_name,
-                ):
-                    await query.answer(text="⛔ You are not authorized to answer this prompt.")
-                    return
-
-                session_key = self._slash_confirm_state.pop(confirm_id, None)
-                if not session_key:
-                    await query.answer(text="This prompt has already been resolved.")
-                    return
-
-                label_map = {
-                    "once": "✅ Approved once",
-                    "always": "🔒 Always approve",
-                    "cancel": "❌ Cancelled",
-                }
-                user_display = getattr(query.from_user, "first_name", "User")
-                label = label_map.get(choice, "Resolved")
-
-                await query.answer(text=label)
-
-                try:
-                    await query.edit_message_text(
-                        text=self.format_message(f"{label} by {user_display}"),
-                        parse_mode=ParseMode.MARKDOWN_V2,
-                        reply_markup=None,
-                    )
-                except Exception:
-                    pass
-
-                # Resolve via the module-level primitive.  The runner stored
-                # a handler keyed by session_key; we run it on the event
-                # loop and (if it returns a string) send it as a follow-up
-                # message in the same chat.
-                try:
-                    from tools import slash_confirm as _slash_confirm_mod
-                    result_text = await _slash_confirm_mod.resolve(
-                        session_key, confirm_id, choice,
-                    )
-                    if result_text and query.message:
-                        # Inherit the prompt message's topic. Supergroup forums
-                        # use message_thread_id; Telegram private DM-topic lanes
-                        # need both the private topic id and the prompt reply anchor.
-                        thread_id = getattr(query.message, "message_thread_id", None)
-                        chat = getattr(query.message, "chat", None)
-                        chat_type = getattr(chat, "type", None)
-                        prompt_message_id = getattr(query.message, "message_id", None)
-                        send_kwargs: Dict[str, Any] = {
-                            "chat_id": int(query.message.chat_id),
-                            "text": self.format_message(result_text),
-                            "parse_mode": ParseMode.MARKDOWN_V2,
-                            **self._link_preview_kwargs(),
-                        }
-                        chat_type_value = getattr(chat_type, "value", chat_type)
-                        is_private_chat = str(chat_type_value).lower() in {
-                            "private",
-                            str(ChatType.PRIVATE).lower(),
-                            str(getattr(ChatType.PRIVATE, "value", ChatType.PRIVATE)).lower(),
-                        }
-                        if thread_id is not None and is_private_chat and prompt_message_id is not None:
-                            reply_to_id = int(prompt_message_id)
-                            send_kwargs["reply_to_message_id"] = reply_to_id
-                            send_kwargs.update(
-                                self._thread_kwargs_for_send(
-                                    str(query.message.chat_id),
-                                    str(thread_id),
-                                    {
-                                        "thread_id": str(thread_id),
-                                        "telegram_dm_topic_reply_fallback": True,
-                                    },
-                                    reply_to_message_id=reply_to_id,
-                                    reply_to_mode=self._reply_to_mode
-                                )
-                            )
-                        elif thread_id is not None:
-                            send_kwargs.update(
-                                self._thread_kwargs_for_send(
-                                    str(query.message.chat_id),
-                                    str(thread_id),
-                                    {"thread_id": str(thread_id)},
-                                    reply_to_mode=self._reply_to_mode
-                                )
-                            )
-                        await self._send_message_with_thread_fallback(**send_kwargs)
-                except Exception as exc:
-                    logger.error("[%s] slash-confirm callback failed: %s", self.name, exc, exc_info=True)
-            return
-
-        # --- Clarify callbacks (cl:clarify_id:idx | cl:clarify_id:other) ---
-        if data.startswith("cl:"):
-            parts = data.split(":", 2)
-            if len(parts) == 3:
-                clarify_id = parts[1]
-                choice_token = parts[2]
-
-                caller_id = str(getattr(query.from_user, "id", ""))
-                if not self._is_callback_user_authorized(
-                    caller_id,
-                    chat_id=query_chat_id,
-                    chat_type=str(query_chat_type) if query_chat_type is not None else None,
-                    thread_id=str(query_thread_id) if query_thread_id is not None else None,
-                    user_name=query_user_name,
-                ):
-                    await query.answer(text="⛔ You are not authorized to answer this prompt.")
-                    return
-
-                session_key = self._clarify_state.get(clarify_id)
-                if not session_key:
-                    await query.answer(text="This prompt has already been resolved.")
-                    return
-
-                user_display = getattr(query.from_user, "first_name", "User")
-
-                if choice_token == "other":
-                    # Flip into text-capture mode and tell the user to type
-                    # their answer.  The gateway's text-intercept will pick
-                    # up the next message in this session and resolve the
-                    # clarify.  Do NOT pop _clarify_state yet — we still
-                    # need it if the user is slow to respond and the entry
-                    # is cleared by something else.
-                    flipped = False
-                    try:
-                        from tools.clarify_gateway import mark_awaiting_text
-                        flipped = mark_awaiting_text(clarify_id)
-                    except Exception as exc:
-                        logger.warning("[%s] mark_awaiting_text failed: %s", self.name, exc)
-
-                    if not flipped:
-                        # Entry evicted (clarify_timeout) or gateway restarted
-                        # between ask and tap — a typed answer would go nowhere.
-                        self._clarify_state.pop(clarify_id, None)
-                        await self._notify_clarify_expired(query, user_display)
-                        return
-
-                    await query.answer(text="✏️ Type your answer in the chat.")
-                    try:
-                        await query.edit_message_text(
-                            text=f"❓ {query.message.text or ''}\n\n<i>Awaiting typed response from {_html.escape(user_display)}…</i>",
-                            parse_mode=ParseMode.HTML,
-                            reply_markup=None,
-                        )
-                    except Exception:
-                        pass
-                    return
-
-                # Numeric choice → resolve immediately with the chosen text
-                try:
-                    idx = int(choice_token)
-                except (ValueError, TypeError):
-                    await query.answer(text="Invalid choice.")
-                    return
-
-                # Look up the choice text from the entry registered in the
-                # clarify primitive.  Fall back to the index if the entry
-                # has been cleaned up (race with timeout / session reset).
-                resolved_text: Optional[str] = None
-                try:
-                    from tools.clarify_gateway import _entries as _clarify_entries  # type: ignore
-                    entry = _clarify_entries.get(clarify_id)
-                    if entry and entry.choices and 0 <= idx < len(entry.choices):
-                        resolved_text = entry.choices[idx]
-                except Exception:
-                    resolved_text = None
-
-                if resolved_text is None:
-                    # Race: entry vanished. Echo the index as a number so
-                    # the agent at least sees an intentional response
-                    # rather than nothing.
-                    resolved_text = f"choice {idx + 1}"
-
-                # Pop state and resolve
-                self._clarify_state.pop(clarify_id, None)
-                try:
-                    from tools.clarify_gateway import resolve_gateway_clarify
-                    resolved = resolve_gateway_clarify(clarify_id, resolved_text)
-                except Exception as exc:
-                    logger.error("[%s] resolve_gateway_clarify failed: %s", self.name, exc)
-                    resolved = False
-
-                if resolved:
-                    await query.answer(text=f"✓ {resolved_text[:60]}")
-                    try:
-                        await query.edit_message_text(
-                            text=f"❓ {_html.escape(query.message.text or '')}\n\n<b>{_html.escape(user_display)}:</b> {_html.escape(resolved_text)}",
-                            parse_mode=ParseMode.HTML,
-                            reply_markup=None,
-                        )
-                    except Exception:
-                        pass
-                    logger.info(
-                        "Telegram clarify button resolved (id=%s, choice=%r, user=%s)",
-                        clarify_id, resolved_text, user_display,
-                    )
-                else:
-                    # Entry evicted (clarify_timeout) or gateway restarted
-                    # between ask and tap — surface this instead of leaving a
-                    # misleading ✓ on a button the agent will never receive.
-                    await self._notify_clarify_expired(query, user_display)
-                    logger.warning(
-                        "Telegram clarify button: resolve_gateway_clarify returned False (id=%s)",
-                        clarify_id,
-                    )
-            return
-
-        # --- Update prompt callbacks ---
-        if not data.startswith("update_prompt:"):
-            return
-        answer = data.split(":", 1)[1]  # "y" or "n"
-        caller_id = str(getattr(query.from_user, "id", ""))
-        if not self._is_callback_user_authorized(
-            caller_id,
-            chat_id=query_chat_id,
-            chat_type=str(query_chat_type) if query_chat_type is not None else None,
-            thread_id=str(query_thread_id) if query_thread_id is not None else None,
-            user_name=query_user_name,
-        ):
-            await query.answer(text="⛔ You are not authorized to answer update prompts.")
-            return
-        await query.answer(text=f"Sent '{answer}' to the update process.")
-        # Edit the message to show the choice and remove buttons
-        label = "Yes" if answer == "y" else "No"
-        try:
-            await query.edit_message_text(
-                text=self.format_message(f"⚕ Update prompt answered: *{label}*"),
-                parse_mode=ParseMode.MARKDOWN_V2,
-                reply_markup=None,
-            )
-        except Exception:
-            pass  # non-fatal if edit fails
-        # Write the response file
-        try:
-            from hermes_constants import get_hermes_home
-            home = get_hermes_home()
-            response_path = home / ".update_response"
-            tmp = response_path.with_suffix(".tmp")
-            tmp.write_text(answer, encoding="utf-8")
-            tmp.replace(response_path)
-            logger.info("Telegram update prompt answered '%s' by user %s",
-                        answer, getattr(query.from_user, "id", "unknown"))
-        except Exception as exc:
-            logger.error("Failed to write update response from callback: %s", exc)
-
     # Maps `gt:<verb>` -> (script-name, extra-args, success-label, is_state).
     # Scripts live in ~/.hermes/scripts/gmail-triage/. `arg` from the callback
     # data is always passed as the first positional arg.
@@ -6706,862 +6225,6 @@ class TelegramAdapter(BasePlatformAdapter):
         "vip":          ("vip-add.sh",         ["email"],  "✓ marked VIP",         True),
         "vip-domain":   ("vip-add.sh",         ["domain"], "✓ marked VIP domain",  True),
     }
-
-    async def _handle_gmail_triage_callback(
-        self,
-        query,
-        data: str,
-        *,
-        query_chat_id,
-        query_chat_type,
-        query_thread_id,
-        query_user_name,
-    ) -> None:
-        """Dispatch a gmail-triage inline-button callback (gt:verb:arg)."""
-        parts = data.split(":", 2)
-        if len(parts) != 3:
-            await query.answer(text="Invalid gmail-triage data.")
-            return
-        verb, arg = parts[1], parts[2]
-
-        caller_id = str(getattr(query.from_user, "id", ""))
-        if not self._is_callback_user_authorized(
-            caller_id,
-            chat_id=query_chat_id,
-            chat_type=str(query_chat_type) if query_chat_type is not None else None,
-            thread_id=str(query_thread_id) if query_thread_id is not None else None,
-            user_name=query_user_name,
-        ):
-            await query.answer(text="⛔ You are not authorized to act on this email.")
-            return
-
-        entry = self._GT_VERB_DISPATCH.get(verb)
-        if not entry:
-            await query.answer(text=f"Unknown verb: {verb}")
-            return
-        script_name, extra_args, success_label, is_state_verb = entry
-
-        script_path = _Path.home() / ".hermes" / "scripts" / "gmail-triage" / script_name
-        if not script_path.exists():
-            await query.answer(text=f"❌ {script_name} missing")
-            logger.error("[%s] gmail-triage script missing: %s", self.name, script_path)
-            return
-
-        cmd = [str(script_path), arg, *extra_args]
-        success = False
-        try:
-            proc = await asyncio.create_subprocess_exec(
-                *cmd,
-                stdout=asyncio.subprocess.PIPE,
-                stderr=asyncio.subprocess.PIPE,
-            )
-            _stdout_bytes, stderr_bytes = await asyncio.wait_for(
-                proc.communicate(), timeout=60,
-            )
-            if proc.returncode == 0:
-                label = success_label
-                success = True
-                logger.info(
-                    "[%s] gmail-triage callback ok: verb=%s arg=%s",
-                    self.name, verb, arg,
-                )
-            else:
-                stderr_text = stderr_bytes.decode("utf-8", errors="replace").strip()
-                last_line = stderr_text.splitlines()[-1] if stderr_text else f"exit {proc.returncode}"
-                label = f"❌ {verb} failed: {last_line[:80]}"
-                logger.error(
-                    "[%s] gmail-triage callback failed: verb=%s arg=%s rc=%s stderr=%s",
-                    self.name, verb, arg, proc.returncode, stderr_text,
-                )
-        except asyncio.TimeoutError:
-            label = f"❌ {verb} timed out"
-            logger.error("[%s] gmail-triage callback timed out: verb=%s arg=%s", self.name, verb, arg)
-        except Exception as exc:
-            label = f"❌ {verb} error: {exc}"
-            logger.error(
-                "[%s] gmail-triage callback exception: verb=%s arg=%s err=%s",
-                self.name, verb, arg, exc, exc_info=True,
-            )
-
-        await query.answer(text=label)
-        if not success:
-            return
-
-        user_display = getattr(query.from_user, "first_name", "User")
-        original_text = (query.message.text or "") if query.message else ""
-        appended = f"{original_text}\n— {label} by {user_display}"
-        try:
-            if is_state_verb:
-                # Sticky state change: append confirmation, KEEP keyboard so
-                # the user can stack further actions on this email.
-                await query.edit_message_text(text=appended)
-            else:
-                # Per-email one-shot: strip keyboard so the action can't fire twice.
-                await query.edit_message_text(text=appended, reply_markup=None)
-        except Exception:
-            pass
-
-    def _missing_media_path_error(self, label: str, path: str) -> str:
-        """Build an actionable file-not-found error for gateway MEDIA delivery.
-
-        Paths like /workspace/... or /output/... often only exist inside the
-        Docker sandbox, while the gateway process runs on the host.
-        """
-        error = f"{label} file not found: {path}"
-        if path.startswith(("/workspace/", "/output/", "/outputs/")):
-            error += (
-                " (path may only exist inside the Docker sandbox. "
-                "Bind-mount a host directory and emit the host-visible "
-                "path in MEDIA: for gateway file delivery.)"
-            )
-        return error
-
-    def _telegram_media_too_large_note(self, label: str, file_size: Any, max_bytes: int) -> str:
-        limit_mb = max(1, max_bytes // (1024 * 1024))
-        try:
-            size_mb = int(file_size or 0) / (1024 * 1024)
-            size_text = f"{size_mb:.1f} MB"
-        except (TypeError, ValueError):
-            size_text = "unknown size"
-        return (
-            f"[Telegram {label} skipped: file size {size_text} exceeds the "
-            f"{limit_mb} MB limit. Ask the user to send a smaller file.]"
-        )
-
-    def _telegram_media_size_allowed(self, source: Any, label: str) -> tuple[bool, Optional[str]]:
-        """Validate Telegram media size before downloading into memory."""
-        max_bytes = int(getattr(self, "_max_doc_bytes", 20 * 1024 * 1024) or 20 * 1024 * 1024)
-        file_size = getattr(source, "file_size", None)
-        try:
-            size = int(file_size or 0)
-        except (TypeError, ValueError):
-            size = 0
-        if size <= 0:
-            return True, None
-        if size <= max_bytes:
-            return True, None
-        return False, self._telegram_media_too_large_note(label, size, max_bytes)
-
-    async def send_voice(
-        self,
-        chat_id: str,
-        audio_path: str,
-        caption: Optional[str] = None,
-        reply_to: Optional[str] = None,
-        metadata: Optional[Dict[str, Any]] = None,
-        **kwargs,
-    ) -> SendResult:
-        """Send audio as a native Telegram voice message or audio file."""
-        if not self._bot:
-            return SendResult(success=False, error="Not connected")
-        
-        try:
-            if not os.path.exists(audio_path):
-                return SendResult(success=False, error=self._missing_media_path_error("Audio", audio_path))
-            
-            # Compute duration locally — Telegram drops it for long clips
-            # (~5 min+), which then show 0:00 in the player.
-            _duration_secs = await asyncio.to_thread(
-                _probe_voice_duration_seconds, audio_path
-            )
-
-            # Render caption markdown (#32029): auto-TTS captions carry the
-            # agent's markdown reply, which showed literal *asterisks* and
-            # [links](...) without a parse_mode. Format to MarkdownV2 when it
-            # fits the 1024-char caption cap; fall back to the raw text
-            # (previous behaviour) when formatting would overflow or the
-            # Bot API rejects the entities.
-            _caption_variants: List[tuple] = []
-            if caption:
-                try:
-                    _formatted_caption = self.format_message(caption)
-                    if utf16_len(_formatted_caption) <= 1024:
-                        _caption_variants.append(
-                            (_formatted_caption, ParseMode.MARKDOWN_V2)
-                        )
-                except Exception:
-                    logger.debug(
-                        "[%s] voice caption MarkdownV2 formatting failed; "
-                        "sending plain caption", self.name, exc_info=True,
-                    )
-                _caption_variants.append((caption[:1024], None))
-            else:
-                _caption_variants.append((None, None))
-
-            with open(audio_path, "rb") as audio_file:
-                ext = os.path.splitext(audio_path)[1].lower()
-                # .ogg / .opus files -> send as voice (round playable bubble)
-                if ext in {".ogg", ".opus"}:
-                    _voice_thread = self._metadata_thread_id(metadata)
-                    reply_to_id = self._reply_to_message_id_for_send(reply_to, metadata, reply_to_mode=self._reply_to_mode)
-                    voice_thread_kwargs = self._thread_kwargs_for_send(
-                        chat_id,
-                        _voice_thread,
-                        metadata,
-                        reply_to_message_id=reply_to_id,
-                        reply_to_mode=self._reply_to_mode
-                    )
-                    msg = None
-                    _last_parse_error: Optional[Exception] = None
-                    for _cap_text, _cap_parse_mode in _caption_variants:
-                        try:
-                            msg = await self._send_with_dm_topic_reply_anchor_retry(
-                                self._bot.send_voice,
-                                {
-                                    "chat_id": normalize_telegram_chat_id(chat_id),
-                                    "voice": audio_file,
-                                    "caption": _cap_text,
-                                    "parse_mode": _cap_parse_mode,
-                                    "reply_to_message_id": reply_to_id,
-                                    "duration": _duration_secs,
-                                    "read_timeout": _MEDIA_SEND_READ_TIMEOUT,
-                                    **voice_thread_kwargs,
-                                    **self._notification_kwargs(metadata),
-                                },
-                                metadata,
-                                reply_to_id,
-                                "voice",
-                                reset_media=lambda: audio_file.seek(0),
-                            )
-                            break
-                        except Exception as _cap_error:
-                            # Only retry the next (plain) variant on entity
-                            # parse failures; anything else is a real send
-                            # error for the outer handler.
-                            if (_cap_parse_mode is not None
-                                    and ("parse" in str(_cap_error).lower()
-                                         or "entit" in str(_cap_error).lower())):
-                                logger.warning(
-                                    "[%s] voice caption MarkdownV2 rejected, "
-                                    "retrying plain: %s",
-                                    self.name,
-                                    _redact_telegram_error_text(_cap_error),
-                                )
-                                _last_parse_error = _cap_error
-                                audio_file.seek(0)
-                                continue
-                            raise
-                    if msg is None:
-                        raise _last_parse_error or RuntimeError(
-                            "Telegram send_voice failed for all caption variants"
-                        )
-                elif ext in {".mp3", ".m4a"}:
-                    # Telegram's Bot API sendAudio only accepts MP3 / M4A.
-                    _audio_thread = self._metadata_thread_id(metadata)
-                    reply_to_id = self._reply_to_message_id_for_send(reply_to, metadata, reply_to_mode=self._reply_to_mode)
-                    audio_thread_kwargs = self._thread_kwargs_for_send(
-                        chat_id,
-                        _audio_thread,
-                        metadata,
-                        reply_to_message_id=reply_to_id,
-                        reply_to_mode=self._reply_to_mode
-                    )
-                    msg = await self._send_with_dm_topic_reply_anchor_retry(
-                        self._bot.send_audio,
-                        {
-                            "chat_id": normalize_telegram_chat_id(chat_id),
-                            "audio": audio_file,
-                            "caption": caption[:1024] if caption else None,
-                            "reply_to_message_id": reply_to_id,
-                            "duration": _duration_secs,
-                            "read_timeout": _MEDIA_SEND_READ_TIMEOUT,
-                            **audio_thread_kwargs,
-                            **self._notification_kwargs(metadata),
-                        },
-                        metadata,
-                        reply_to_id,
-                        "audio",
-                        reset_media=lambda: audio_file.seek(0),
-                    )
-                else:
-                    # Formats Telegram can't play natively (.wav, .flac, ...)
-                    # — fall back to document delivery instead of raising.
-                    return await self.send_document(
-                        chat_id=chat_id,
-                        file_path=audio_path,
-                        caption=caption,
-                        reply_to=reply_to,
-                        metadata=metadata,
-                    )
-            return SendResult(success=True, message_id=str(msg.message_id))
-        except Exception as e:
-            logger.error(
-                "[%s] Failed to send Telegram voice/audio, falling back to base adapter: %s",
-                self.name,
-                _redact_telegram_error_text(e),
-                exc_info=True,
-            )
-            return await super().send_voice(chat_id, audio_path, caption, reply_to, metadata=metadata)
-
-    async def send_multiple_images(
-        self,
-        chat_id: str,
-        images: List[tuple],
-        metadata: Optional[Dict[str, Any]] = None,
-        human_delay: float = 0.0,
-    ) -> None:
-        """Send a batch of images natively via Telegram's media group API.
-
-        Telegram's ``send_media_group`` bundles up to 10 photos/videos into
-        a single album. Larger batches are chunked. Animated GIFs cannot
-        go into a media group (they require ``send_animation``), so they
-        are peeled off and sent individually via the base default path.
-
-        URL-based photos go into the group directly; local files are
-        opened as byte streams. On failure the whole batch falls back to
-        the base adapter's per-image loop.
-        """
-        if not self._bot:
-            return
-        if not images:
-            return
-
-        try:
-            from telegram import InputMediaPhoto
-        except Exception as exc:  # pragma: no cover - missing SDK
-            logger.warning(
-                "[%s] InputMediaPhoto unavailable, falling back to per-image send: %s",
-                self.name, exc,
-            )
-            await super().send_multiple_images(chat_id, images, metadata, human_delay)
-            return
-
-        # Peel off animations — they need send_animation, not send_media_group
-        animations: List[tuple] = []
-        photos: List[tuple] = []
-        for image_url, alt_text in images:
-            if not image_url.startswith("file://") and self._is_animation_url(image_url):
-                animations.append((image_url, alt_text))
-            else:
-                photos.append((image_url, alt_text))
-
-        # Animations: route through the base default (per-image send_animation)
-        if animations:
-            await super().send_multiple_images(
-                chat_id, animations, metadata, human_delay=human_delay,
-            )
-
-        if not photos:
-            return
-
-        from urllib.parse import unquote as _unquote
-        _thread = self._metadata_thread_id(metadata)
-
-        # Chunk into groups of 10 (Telegram's album limit)
-        CHUNK = 10
-        chunks = [photos[i:i + CHUNK] for i in range(0, len(photos), CHUNK)]
-
-        for chunk_idx, chunk in enumerate(chunks):
-            if human_delay > 0 and chunk_idx > 0:
-                await asyncio.sleep(human_delay)
-
-            media: List[Any] = []
-            opened_files: List[Any] = []
-            try:
-                for image_url, alt_text in chunk:
-                    caption = alt_text[:1024] if alt_text else None
-                    if image_url.startswith("file://"):
-                        local_path = _unquote(image_url[7:])
-                        if not os.path.exists(local_path):
-                            logger.warning(
-                                "[%s] Skipping missing image in media group: %s",
-                                self.name, local_path,
-                            )
-                            continue
-                        fh = open(local_path, "rb")
-                        opened_files.append(fh)
-                        media.append(InputMediaPhoto(media=fh, caption=caption))
-                    else:
-                        media.append(InputMediaPhoto(media=image_url, caption=caption))
-
-                if not media:
-                    continue
-
-                logger.info(
-                    "[%s] Sending media group of %d photo(s) (chunk %d/%d)",
-                    self.name, len(media), chunk_idx + 1, len(chunks),
-                )
-                reply_to_id = self._reply_to_message_id_for_send(None, metadata, reply_to_mode=self._reply_to_mode)
-                thread_kwargs = self._thread_kwargs_for_send(
-                    chat_id,
-                    _thread,
-                    metadata,
-                    reply_to_message_id=reply_to_id,
-                    reply_to_mode=self._reply_to_mode
-                )
-
-                def _reset_opened_files() -> None:
-                    for fh in opened_files:
-                        try:
-                            fh.seek(0)
-                        except Exception:
-                            pass
-
-                await self._send_with_dm_topic_reply_anchor_retry(
-                    self._bot.send_media_group,
-                    {
-                        "chat_id": normalize_telegram_chat_id(chat_id),
-                        "media": media,
-                        "reply_to_message_id": reply_to_id,
-                        "read_timeout": _MEDIA_SEND_READ_TIMEOUT,
-                        **thread_kwargs,
-                        **self._notification_kwargs(metadata),
-                    },
-                    metadata,
-                    reply_to_id,
-                    "media group",
-                    reset_media=_reset_opened_files,
-                )
-            except Exception as e:
-                logger.warning(
-                    "[%s] send_media_group failed (chunk %d/%d), falling back to per-image: %s",
-                    self.name, chunk_idx + 1, len(chunks), _redact_telegram_error_text(e),
-                    exc_info=True,
-                )
-                # Fallback: send each photo in this chunk individually
-                await super().send_multiple_images(
-                    chat_id, chunk, metadata, human_delay=human_delay,
-                )
-            finally:
-                for fh in opened_files:
-                    try:
-                        fh.close()
-                    except Exception:
-                        pass
-
-    async def send_image_file(
-        self,
-        chat_id: str,
-        image_path: str,
-        caption: Optional[str] = None,
-        reply_to: Optional[str] = None,
-        metadata: Optional[Dict[str, Any]] = None,
-        **kwargs,
-    ) -> SendResult:
-        """Send a local image file natively as a Telegram photo."""
-        if not self._bot:
-            return SendResult(success=False, error="Not connected")
-
-        try:
-            if not os.path.exists(image_path):
-                return SendResult(success=False, error=self._missing_media_path_error("Image", image_path))
-
-            _thread = self._metadata_thread_id(metadata)
-            reply_to_id = self._reply_to_message_id_for_send(reply_to, metadata, reply_to_mode=self._reply_to_mode)
-            thread_kwargs = self._thread_kwargs_for_send(
-                chat_id,
-                _thread,
-                metadata,
-                reply_to_message_id=reply_to_id,
-                reply_to_mode=self._reply_to_mode
-            )
-            with open(image_path, "rb") as image_file:
-                msg = await self._send_with_dm_topic_reply_anchor_retry(
-                    self._bot.send_photo,
-                    {
-                        "chat_id": normalize_telegram_chat_id(chat_id),
-                        "photo": image_file,
-                        "caption": caption[:1024] if caption else None,
-                        "reply_to_message_id": reply_to_id,
-                        "read_timeout": _MEDIA_SEND_READ_TIMEOUT,
-                        **thread_kwargs,
-                        **self._notification_kwargs(metadata),
-                    },
-                    metadata,
-                    reply_to_id,
-                    "photo",
-                    reset_media=lambda: image_file.seek(0),
-                )
-            return SendResult(success=True, message_id=str(msg.message_id))
-        except Exception as e:
-            error_str = str(e)
-            # Dimension-related errors are the expected case for valid image
-            # files that Telegram just refuses as photos (screenshots, extreme
-            # aspect ratios). Log at INFO because the document fallback is
-            # the correct path. Any other send_photo failure also falls back
-            # to document (rate limits, corrupt file markers, format edge
-            # cases), but at WARNING because it's unexpected and worth
-            # surfacing in logs.
-            is_dim_error = (
-                "Photo_invalid_dimensions" in error_str
-                or "PHOTO_INVALID_DIMENSIONS" in error_str
-            )
-            if is_dim_error:
-                logger.info(
-                    "[%s] Image dimensions exceed Telegram photo limits, "
-                    "sending as document: %s",
-                    self.name,
-                    image_path,
-                )
-            else:
-                logger.warning(
-                    "[%s] Failed to send Telegram local image as photo, "
-                    "trying document fallback: %s",
-                    self.name,
-                    _redact_telegram_error_text(e),
-                    exc_info=True,
-                )
-            # Fallback to sending as document (file) — no dimension limit,
-            # only 50MB size limit. If even that fails, fall back to the
-            # base adapter's text-only "Image: /path" rendering.
-            try:
-                return await self.send_document(
-                    chat_id=chat_id,
-                    file_path=image_path,
-                    caption=caption,
-                    file_name=os.path.basename(image_path),
-                    reply_to=reply_to,
-                    metadata=metadata,
-                )
-            except Exception as doc_err:
-                logger.error(
-                    "[%s] Failed to send Telegram local image as document, "
-                    "falling back to base adapter: %s",
-                    self.name,
-                    doc_err,
-                    exc_info=True,
-                )
-                return await super().send_image_file(chat_id, image_path, caption, reply_to, metadata=metadata)
-
-    async def send_document(
-        self,
-        chat_id: str,
-        file_path: str,
-        caption: Optional[str] = None,
-        file_name: Optional[str] = None,
-        reply_to: Optional[str] = None,
-        metadata: Optional[Dict[str, Any]] = None,
-        **kwargs,
-    ) -> SendResult:
-        """Send a document/file natively as a Telegram file attachment."""
-        if not self._bot:
-            return SendResult(success=False, error="Not connected")
-
-        try:
-            if not os.path.exists(file_path):
-                return SendResult(success=False, error=self._missing_media_path_error("File", file_path))
-
-            display_name = file_name or os.path.basename(file_path)
-            _thread = self._metadata_thread_id(metadata)
-            reply_to_id = self._reply_to_message_id_for_send(reply_to, metadata, reply_to_mode=self._reply_to_mode)
-            thread_kwargs = self._thread_kwargs_for_send(
-                chat_id,
-                _thread,
-                metadata,
-                reply_to_message_id=reply_to_id,
-                reply_to_mode=self._reply_to_mode
-            )
-
-            with open(file_path, "rb") as f:
-                msg = await self._send_with_dm_topic_reply_anchor_retry(
-                    self._bot.send_document,
-                    {
-                        "chat_id": normalize_telegram_chat_id(chat_id),
-                        "document": f,
-                        "filename": display_name,
-                        "caption": caption[:1024] if caption else None,
-                        "reply_to_message_id": reply_to_id,
-                        "read_timeout": _MEDIA_SEND_READ_TIMEOUT,
-                        **thread_kwargs,
-                        **self._notification_kwargs(metadata),
-                    },
-                    metadata,
-                    reply_to_id,
-                    "document",
-                    reset_media=lambda: f.seek(0),
-                )
-            return SendResult(success=True, message_id=str(msg.message_id))
-        except Exception as e:
-            logger.warning(
-                "[%s] Failed to send document: %s",
-                self.name, _redact_telegram_error_text(e),
-            )
-            return await super().send_document(chat_id, file_path, caption, file_name, reply_to, metadata=metadata)
-
-    async def send_video(
-        self,
-        chat_id: str,
-        video_path: str,
-        caption: Optional[str] = None,
-        reply_to: Optional[str] = None,
-        metadata: Optional[Dict[str, Any]] = None,
-        **kwargs,
-    ) -> SendResult:
-        """Send a video natively as a Telegram video message."""
-        if not self._bot:
-            return SendResult(success=False, error="Not connected")
-
-        try:
-            if not os.path.exists(video_path):
-                return SendResult(success=False, error=self._missing_media_path_error("Video", video_path))
-
-            _thread = self._metadata_thread_id(metadata)
-            reply_to_id = self._reply_to_message_id_for_send(reply_to, metadata, reply_to_mode=self._reply_to_mode)
-            thread_kwargs = self._thread_kwargs_for_send(
-                chat_id,
-                _thread,
-                metadata,
-                reply_to_message_id=reply_to_id,
-                reply_to_mode=self._reply_to_mode
-            )
-            with open(video_path, "rb") as f:
-                msg = await self._send_with_dm_topic_reply_anchor_retry(
-                    self._bot.send_video,
-                    {
-                        "chat_id": normalize_telegram_chat_id(chat_id),
-                        "video": f,
-                        "caption": caption[:1024] if caption else None,
-                        "reply_to_message_id": reply_to_id,
-                        "read_timeout": _MEDIA_SEND_READ_TIMEOUT,
-                        **thread_kwargs,
-                        **self._notification_kwargs(metadata),
-                    },
-                    metadata,
-                    reply_to_id,
-                    "video",
-                    reset_media=lambda: f.seek(0),
-                )
-            return SendResult(success=True, message_id=str(msg.message_id))
-        except Exception as e:
-            logger.warning(
-                "[%s] Failed to send video: %s",
-                self.name, _redact_telegram_error_text(e),
-            )
-            return await super().send_video(chat_id, video_path, caption, reply_to, metadata=metadata)
-
-    async def send_image(
-        self,
-        chat_id: str,
-        image_url: str,
-        caption: Optional[str] = None,
-        reply_to: Optional[str] = None,
-        metadata: Optional[Dict[str, Any]] = None,
-    ) -> SendResult:
-        """Send an image natively as a Telegram photo.
-        
-        Tries URL-based send first (fast, works for <5MB images).
-        Falls back to downloading and uploading as file (supports up to 10MB).
-        """
-        if not self._bot:
-            return SendResult(success=False, error="Not connected")
-
-        from tools.url_safety import is_safe_url
-        if not is_safe_url(image_url):
-            logger.warning("[%s] Blocked unsafe image URL (SSRF protection)", self.name)
-            return await super().send_image(chat_id, image_url, caption, reply_to, metadata=metadata)
-
-        try:
-            # Telegram can send photos directly from URLs (up to ~5MB)
-            _photo_thread = self._metadata_thread_id(metadata)
-            reply_to_id = self._reply_to_message_id_for_send(reply_to, metadata, reply_to_mode=self._reply_to_mode)
-            photo_thread_kwargs = self._thread_kwargs_for_send(
-                chat_id,
-                _photo_thread,
-                metadata,
-                reply_to_message_id=reply_to_id,
-                reply_to_mode=self._reply_to_mode
-            )
-            msg = await self._send_with_dm_topic_reply_anchor_retry(
-                self._bot.send_photo,
-                {
-                    "chat_id": normalize_telegram_chat_id(chat_id),
-                    "photo": image_url,
-                    "caption": caption[:1024] if caption else None,
-                    "reply_to_message_id": reply_to_id,
-                    "read_timeout": _MEDIA_SEND_READ_TIMEOUT,
-                    **photo_thread_kwargs,
-                    **self._notification_kwargs(metadata),
-                },
-                metadata,
-                reply_to_id,
-                "URL photo",
-            )
-            return SendResult(success=True, message_id=str(msg.message_id))
-        except Exception as e:
-            logger.warning(
-                "[%s] URL-based send_photo failed, trying file upload: %s",
-                self.name,
-                _redact_telegram_error_text(e),
-                exc_info=True,
-            )
-            # Fallback: download and upload as file (supports up to 10MB)
-            try:
-                from gateway.platforms.base import _ssrf_redirect_guard
-                from tools.url_safety import create_ssrf_safe_async_client
-
-                async with create_ssrf_safe_async_client(
-                    timeout=30.0,
-                    event_hooks={"response": [_ssrf_redirect_guard]},
-                ) as client:
-                    resp = await client.get(image_url)
-                    resp.raise_for_status()
-                    image_data = resp.content
-
-                upload_thread_kwargs = self._thread_kwargs_for_send(
-                    chat_id,
-                    _photo_thread,
-                    metadata,
-                    reply_to_message_id=reply_to_id,
-                    reply_to_mode=self._reply_to_mode
-                )
-                msg = await self._send_with_dm_topic_reply_anchor_retry(
-                    self._bot.send_photo,
-                    {
-                        "chat_id": normalize_telegram_chat_id(chat_id),
-                        "photo": image_data,
-                        "caption": caption[:1024] if caption else None,
-                        "reply_to_message_id": reply_to_id,
-                        "read_timeout": _MEDIA_SEND_READ_TIMEOUT,
-                        **upload_thread_kwargs,
-                        **self._notification_kwargs(metadata),
-                    },
-                    metadata,
-                    reply_to_id,
-                    "uploaded photo",
-                )
-                return SendResult(success=True, message_id=str(msg.message_id))
-            except Exception as e2:
-                logger.error(
-                    "[%s] File upload send_photo also failed: %s",
-                    self.name,
-                    e2,
-                    exc_info=True,
-                )
-                # Final fallback: send URL as text
-                return await super().send_image(chat_id, image_url, caption, reply_to, metadata=metadata)
-
-    async def send_animation(
-        self,
-        chat_id: str,
-        animation_url: str,
-        caption: Optional[str] = None,
-        reply_to: Optional[str] = None,
-        metadata: Optional[Dict[str, Any]] = None,
-    ) -> SendResult:
-        """Send an animated GIF natively as a Telegram animation (auto-plays inline)."""
-        if not self._bot:
-            return SendResult(success=False, error="Not connected")
-        
-        try:
-            _anim_thread = self._metadata_thread_id(metadata)
-            reply_to_id = self._reply_to_message_id_for_send(reply_to, metadata, reply_to_mode=self._reply_to_mode)
-            animation_thread_kwargs = self._thread_kwargs_for_send(
-                chat_id,
-                _anim_thread,
-                metadata,
-                reply_to_message_id=reply_to_id,
-                reply_to_mode=self._reply_to_mode
-            )
-            msg = await self._send_with_dm_topic_reply_anchor_retry(
-                self._bot.send_animation,
-                {
-                    "chat_id": normalize_telegram_chat_id(chat_id),
-                    "animation": animation_url,
-                    "caption": caption[:1024] if caption else None,
-                    "reply_to_message_id": reply_to_id,
-                    "read_timeout": _MEDIA_SEND_READ_TIMEOUT,
-                    **animation_thread_kwargs,
-                    **self._notification_kwargs(metadata),
-                },
-                metadata,
-                reply_to_id,
-                "animation",
-            )
-            return SendResult(success=True, message_id=str(msg.message_id))
-        except Exception as e:
-            logger.error(
-                "[%s] Failed to send Telegram animation, falling back to photo: %s",
-                self.name,
-                _redact_telegram_error_text(e),
-                exc_info=True,
-            )
-            # Fallback: try as a regular photo
-            return await self.send_image(chat_id, animation_url, caption, reply_to, metadata=metadata)
-
-    @staticmethod
-    def _is_transient_typing_error(exc: Exception) -> bool:
-        """Return True for Telegram typing errors worth cooling down."""
-        retry_after = getattr(exc, "retry_after", None)
-        if retry_after is not None:
-            return True
-
-        status_code = getattr(exc, "status_code", None) or getattr(exc, "code", None)
-        if isinstance(status_code, int) and (status_code == 429 or status_code >= 500):
-            return True
-
-        text = str(exc).lower()
-        if any(marker in text for marker in ("too many requests", "rate limit", "timed out", "timeout", "temporar")):
-            return True
-        if isinstance(exc, (OSError, TimeoutError, ConnectionError, asyncio.TimeoutError)):
-            return True
-        return False
-
-    def _record_typing_cooldown(self, chat_id: str, exc: Exception) -> None:
-        """Suppress Telegram typing refreshes for this chat after transient failures."""
-        if not hasattr(self, "_telegram_typing_cooldown_until"):
-            self._telegram_typing_cooldown_until = {}
-        loop = asyncio.get_running_loop()
-        retry_after = getattr(exc, "retry_after", None)
-        try:
-            delay = float(retry_after) if retry_after is not None else self._telegram_typing_cooldown_seconds
-        except (TypeError, ValueError):
-            delay = self._telegram_typing_cooldown_seconds
-        delay = max(1.0, min(delay, 300.0))
-        self._telegram_typing_cooldown_until[str(chat_id)] = loop.time() + delay
-
-    def _typing_in_cooldown(self, chat_id: str) -> bool:
-        if not hasattr(self, "_telegram_typing_cooldown_until"):
-            self._telegram_typing_cooldown_until = {}
-            self._telegram_typing_cooldown_seconds = 30.0
-        until = self._telegram_typing_cooldown_until.get(str(chat_id))
-        if until is None:
-            return False
-        if asyncio.get_running_loop().time() < until:
-            return True
-        self._telegram_typing_cooldown_until.pop(str(chat_id), None)
-        return False
-
-    async def send_typing(self, chat_id: str, metadata: Optional[Dict[str, Any]] = None) -> None:
-        """Send typing indicator."""
-        if not self._bot or self._typing_in_cooldown(chat_id):
-            return
-
-        _is_dm_topic: bool = False
-        message_thread_id: Optional[int] = None
-        try:
-            _typing_thread = self._metadata_thread_id(metadata)
-            _is_dm_topic = bool(metadata and metadata.get("telegram_dm_topic_reply_fallback"))
-            message_thread_id = self._message_thread_id_for_typing(_typing_thread)
-            await self._bot.send_chat_action(
-                chat_id=normalize_telegram_chat_id(chat_id),
-                action="typing",
-                message_thread_id=message_thread_id,
-            )
-            self._telegram_typing_cooldown_until.pop(str(chat_id), None)
-        except Exception as e:
-            # For DM topic lanes, Telegram may reject message_thread_id.
-            # Fall back to sending typing without thread_id so the typing
-            # indicator at least appears in the main DM view.
-            if _is_dm_topic and message_thread_id is not None:
-                try:
-                    await self._bot.send_chat_action(
-                        chat_id=normalize_telegram_chat_id(chat_id),
-                        action="typing",
-                    )
-                    self._telegram_typing_cooldown_until.pop(str(chat_id), None)
-                    return
-                except Exception as fallback_exc:
-                    if self._is_transient_typing_error(fallback_exc):
-                        self._record_typing_cooldown(chat_id, fallback_exc)
-            elif self._is_transient_typing_error(e):
-                self._record_typing_cooldown(chat_id, e)
-            # Typing failures are non-fatal; log at debug level only.
-            logger.debug(
-                "[%s] Failed to send Telegram typing indicator: %s",
-                self.name,
-                _redact_telegram_error_text(e),
-                exc_info=True,
-            )
 
     async def get_chat_info(self, chat_id: str) -> Dict[str, Any]:
         """Get information about a Telegram chat."""
@@ -7596,258 +6259,6 @@ class TelegramAdapter(BasePlatformAdapter):
                 exc_info=True,
             )
             return {"name": str(chat_id), "type": "dm", "error": str(e)}
-
-    def format_message(self, content: str) -> str:
-        """
-        Convert standard markdown to Telegram MarkdownV2 format.
-
-        Protected regions (code blocks, inline code) are extracted first so
-        their contents are never modified.  Standard markdown constructs
-        (headers, bold, italic, links) are translated to MarkdownV2 syntax,
-        and all remaining special characters are escaped.
-        """
-        if not content:
-            return content
-
-        placeholders: dict = {}
-        counter = [0]
-
-        def _ph(value: str) -> str:
-            """Stash *value* behind a placeholder token that survives escaping."""
-            key = f"\x00PH{counter[0]}\x00"
-            counter[0] += 1
-            placeholders[key] = value
-            return key
-
-        text = content
-
-        # 0) Rewrite GFM-style pipe tables into Telegram-friendly row groups
-        #    before the normal MarkdownV2 conversions run.
-        text = _wrap_markdown_tables(text)
-
-        # 1) Protect fenced code blocks (``` ... ```)
-        #    Per MarkdownV2 spec, \ and ` inside pre/code must be escaped.
-        def _protect_fenced(m):
-            raw = m.group(0)
-            # Split off opening ``` (with optional language) and closing ```
-            open_end = raw.index('\n') + 1 if '\n' in raw[3:] else 3
-            opening = raw[:open_end]
-            body_and_close = raw[open_end:]
-            body = body_and_close[:-3]
-            body = body.replace('\\', '\\\\').replace('`', '\\`')
-            return _ph(opening + body + '```')
-
-        text = re.sub(
-            r'(```(?:[^\n]*\n)?[\s\S]*?```)',
-            _protect_fenced,
-            text,
-        )
-
-        # 2) Protect inline code (`...`)
-        #    Escape \ inside inline code per MarkdownV2 spec.
-        text = re.sub(
-            r'(`[^`]+`)',
-            lambda m: _ph(m.group(0).replace('\\', '\\\\')),
-            text,
-        )
-
-        # 3) Convert markdown links – escape the display text; inside the URL
-        #    only ')' and '\' need escaping per the MarkdownV2 spec.
-        def _convert_link(m):
-            display = _escape_mdv2(m.group(1))
-            url = m.group(2).replace('\\', '\\\\').replace(')', '\\)')
-            return _ph(f'[{display}]({url})')
-
-        text = re.sub(r'\[([^\]]+)\]\(([^()]*(?:\([^()]*\)[^()]*)*)\)', _convert_link, text)
-
-        # 4) Convert markdown headers (## Title) → bold *Title*
-        def _convert_header(m):
-            inner = m.group(1).strip()
-            # Strip redundant bold markers that may appear inside a header
-            inner = re.sub(r'\*\*(.+?)\*\*', r'\1', inner)
-            return _ph(f'*{_escape_mdv2(inner)}*')
-
-        text = re.sub(
-            r'^#{1,6}\s+(.+)$', _convert_header, text, flags=re.MULTILINE
-        )
-
-        # 5) Convert bold: **text** → *text* (MarkdownV2 bold)
-        text = re.sub(
-            r'\*\*(.+?)\*\*',
-            lambda m: _ph(f'*{_escape_mdv2(m.group(1))}*'),
-            text,
-        )
-
-        # 6) Convert italic: *text* (single asterisk) → _text_ (MarkdownV2 italic)
-        #    [^*\n]+ prevents matching across newlines (which would corrupt
-        #    bullet lists using * markers and multi-line content).
-        text = re.sub(
-            r'\*([^*\n]+)\*',
-            lambda m: _ph(f'_{_escape_mdv2(m.group(1))}_'),
-            text,
-        )
-
-        # 7) Convert strikethrough: ~~text~~ → ~text~ (MarkdownV2)
-        text = re.sub(
-            r'~~(.+?)~~',
-            lambda m: _ph(f'~{_escape_mdv2(m.group(1))}~'),
-            text,
-        )
-
-        # 8) Convert spoiler: ||text|| → ||text|| (protect from | escaping)
-        text = re.sub(
-            r'\|\|(.+?)\|\|',
-            lambda m: _ph(f'||{_escape_mdv2(m.group(1))}||'),
-            text,
-        )
-
-        # 9) Convert blockquotes: > at line start → protect > from escaping
-        #    Handle both regular blockquotes (> text) and expandable blockquotes
-        #    (Telegram MarkdownV2: **> for expandable start, || to end the quote)
-        def _convert_blockquote(m):
-            prefix = m.group(1)  # >, >>, >>>, **>, or **>> etc.
-            content = m.group(2)
-            # Check if content ends with || (expandable blockquote end marker)
-            # In this case, preserve the trailing || unescaped for Telegram
-            if prefix.startswith('**') and content.endswith('||'):
-                return _ph(f'{prefix} {_escape_mdv2(content[:-2])}||')
-            return _ph(f'{prefix} {_escape_mdv2(content)}')
-
-        text = re.sub(
-            r'^((?:\*\*)?>{1,3}) (.+)$',
-            _convert_blockquote,
-            text,
-            flags=re.MULTILINE,
-        )
-
-        # 10) Escape remaining special characters in plain text
-        text = _escape_mdv2(text)
-
-        # 11) Restore placeholders in reverse insertion order so that
-        #    nested references (a placeholder inside another) resolve correctly.
-        for key in reversed(list(placeholders.keys())):
-            text = text.replace(key, placeholders[key])
-
-        # 12) Safety net: escape unescaped ( ) { } that slipped through
-        #     placeholder processing.  Split the text into code/non-code
-        #     segments so we never touch content inside ``` or ` spans.
-        _code_split = re.split(r'(```[\s\S]*?```|`[^`]+`)', text)
-        _safe_parts = []
-        for _idx, _seg in enumerate(_code_split):
-            if _idx % 2 == 1:
-                # Inside code span/block — leave untouched
-                _safe_parts.append(_seg)
-            else:
-                # Outside code — escape bare ( ) { }
-                def _esc_bare(m, _seg=_seg):
-                    s = m.start()
-                    ch = m.group(0)
-                    # Already escaped
-                    if s > 0 and _seg[s - 1] == '\\':
-                        return ch
-                    # ( that opens a MarkdownV2 link [text](url)
-                    if ch == '(' and s > 0 and _seg[s - 1] == ']':
-                        return ch
-                    # ) that closes a link URL
-                    if ch == ')':
-                        before = _seg[:s]
-                        if '](http' in before or '](' in before:
-                            # Check depth
-                            depth = 0
-                            for j in range(s - 1, max(s - 2000, -1), -1):
-                                if _seg[j] == '(':
-                                    depth -= 1
-                                    if depth < 0:
-                                        if j > 0 and _seg[j - 1] == ']':
-                                            return ch
-                                        break
-                                elif _seg[j] == ')':
-                                    depth += 1
-                    return '\\' + ch
-                _safe_parts.append(re.sub(r'[(){}]', _esc_bare, _seg))
-        text = ''.join(_safe_parts)
-
-        return text
-
-    # ── Group mention gating ──────────────────────────────────────────────
-
-    def _telegram_require_mention(self) -> bool:
-        """Return whether group chats should require an explicit bot trigger."""
-        configured = self.config.extra.get("require_mention")
-        if configured is not None:
-            if isinstance(configured, str):
-                return configured.lower() in {"true", "1", "yes", "on"}
-            return bool(configured)
-        return os.getenv("TELEGRAM_REQUIRE_MENTION", "false").lower() in {"true", "1", "yes", "on"}
-
-    def _telegram_observe_unmentioned_group_messages(self) -> bool:
-        """Return whether skipped unmentioned group messages are stored as context.
-
-        When enabled with ``require_mention``, Telegram matches the Yuanbao /
-        OpenClaw-style group UX: observe ordinary group chatter in the session
-        transcript, but only dispatch the agent when the bot is explicitly
-        addressed.
-        """
-        configured = self.config.extra.get("observe_unmentioned_group_messages")
-        if configured is None:
-            configured = self.config.extra.get("ingest_unmentioned_group_messages")
-        if configured is not None:
-            if isinstance(configured, str):
-                return configured.lower() in {"true", "1", "yes", "on"}
-            return bool(configured)
-        return os.getenv("TELEGRAM_OBSERVE_UNMENTIONED_GROUP_MESSAGES", "false").lower() in {"true", "1", "yes", "on"}
-
-    def _telegram_guest_mode(self) -> bool:
-        """Return whether non-allowlisted groups may trigger via direct @mention."""
-        configured = self.config.extra.get("guest_mode")
-        if configured is not None:
-            if isinstance(configured, str):
-                return configured.lower() in {"true", "1", "yes", "on"}
-            return bool(configured)
-        return os.getenv("TELEGRAM_GUEST_MODE", "false").lower() in {"true", "1", "yes", "on"}
-
-    def _telegram_exclusive_bot_mentions(self) -> bool:
-        """Return whether explicit @...bot mentions exclusively route group messages."""
-        configured = self.config.extra.get("exclusive_bot_mentions")
-        if configured is not None:
-            if isinstance(configured, str):
-                return configured.lower() in {"true", "1", "yes", "on"}
-            return bool(configured)
-        return os.getenv("TELEGRAM_EXCLUSIVE_BOT_MENTIONS", "true").lower() in {"true", "1", "yes", "on"}
-
-    def _telegram_free_response_chats(self) -> set[str]:
-        raw = self.config.extra.get("free_response_chats")
-        if raw is None:
-            raw = _scoped_gate_env("TELEGRAM_FREE_RESPONSE_CHATS")
-        if isinstance(raw, list):
-            return {str(part).strip() for part in raw if str(part).strip()}
-        return {part.strip() for part in str(raw).split(",") if part.strip()}
-
-    def _telegram_free_response_topics(self) -> set[str]:
-        """Return topic-level free-response allowlist entries as ``<chat_id>:<thread_id>``.
-
-        Unlike ``free_response_chats`` (whole-chat), each entry opens a single
-        forum topic for free-response. A missing/omitted thread id on incoming
-        messages is normalized to the General topic (``1``).
-        """
-        raw = self.config.extra.get("free_response_topics")
-        if raw is None:
-            raw = _scoped_gate_env("TELEGRAM_FREE_RESPONSE_TOPICS")
-        if isinstance(raw, list):
-            return {str(part).strip() for part in raw if str(part).strip()}
-        return {part.strip() for part in str(raw).split(",") if part.strip()}
-
-    def _telegram_is_free_response_topic(self, message: Message) -> bool:
-        """True when the message's chat/topic pair is in ``free_response_topics``."""
-        topics = self._telegram_free_response_topics()
-        if not topics:
-            return False
-        chat_id = str(getattr(getattr(message, "chat", None), "id", ""))
-        if not chat_id:
-            return False
-        thread_id = self._effective_message_thread_id(message)
-        topic_id = str(thread_id) if thread_id is not None else self._GENERAL_TOPIC_THREAD_ID
-        return f"{chat_id}:{topic_id}" in topics
 
     def _telegram_allowed_chats(self) -> set[str]:
         """Return the whitelist of group/supergroup chat IDs the bot will respond in.
@@ -7924,68 +6335,6 @@ class TelegramAdapter(BasePlatformAdapter):
                 logger.warning("[%s] Ignoring invalid Telegram thread id: %r", self.name, value)
         return ignored
 
-    def _compile_mention_patterns(self) -> List[re.Pattern]:
-        """Compile optional regex wake-word patterns for group triggers."""
-        patterns = self.config.extra.get("mention_patterns")
-        if patterns is None:
-            raw = os.getenv("TELEGRAM_MENTION_PATTERNS", "").strip()
-            if raw:
-                try:
-                    loaded = json.loads(raw)
-                except Exception:
-                    loaded = [part.strip() for part in raw.splitlines() if part.strip()]
-                    if not loaded:
-                        loaded = [part.strip() for part in raw.split(",") if part.strip()]
-                patterns = loaded
-
-        if patterns is None:
-            # Parity with the historical inline implementation: return before
-            # evaluating ``self.name`` (tests construct bare adapters via
-            # object.__new__ that lack the attributes ``name`` reads).
-            return []
-
-        return compile_mention_patterns(
-            patterns,
-            log_prefix=self.name,
-            platform_label="telegram",
-            display_label="Telegram",
-            logger_=logger,
-        )
-
-    def _is_group_chat(self, message: Message) -> bool:
-        chat = getattr(message, "chat", None)
-        if not chat:
-            return False
-        chat_type = str(getattr(chat, "type", "")).split(".")[-1].lower()
-        return chat_type in {"group", "supergroup"}
-
-    @classmethod
-    def _effective_message_thread_id(cls, message: Message) -> Optional[str]:
-        """Return the routable thread id for a Telegram message.
-
-        Forum supergroup messages posted in the General topic arrive with
-        ``message_thread_id=None`` while Telegram itself addresses that topic
-        as thread id ``1``.  Ordinary replies are the opposite footgun:
-        Telegram populates ``message_thread_id`` with a reply-UI anchor id on
-        plain group/DM replies, but those ids are not topic/session routing
-        ids and must not be treated as such.  Gating, skill binding, and
-        outbound routing must all agree on the same normalized value.
-        """
-        chat = getattr(message, "chat", None)
-        chat_type = str(getattr(chat, "type", "")).split(".")[-1].lower() if chat else ""
-        raw = getattr(message, "message_thread_id", None)
-        is_topic_message = bool(getattr(message, "is_topic_message", False))
-        is_forum_group = chat_type in ("group", "supergroup") and getattr(chat, "is_forum", False) is True
-        if raw is not None:
-            if is_forum_group or (chat_type in ("group", "supergroup") and is_topic_message):
-                return str(raw)
-            if chat_type == "private" and is_topic_message:
-                return str(raw)
-            return None
-        if is_forum_group:
-            return cls._GENERAL_TOPIC_THREAD_ID
-        return None
-
     # Telegram bot handles historically had to end in "bot", but collectible
     # (Fragment) usernames can be assigned to bots and drop that suffix
     # entirely (@jarvis, @pic, ...). This pattern is used ONLY to decide
@@ -7995,176 +6344,7 @@ class TelegramAdapter(BasePlatformAdapter):
     # How long an observed identity is trusted before the heartbeat re-checks.
     _BOT_IDENTITY_TTL_SECONDS = 300.0
 
-    def _current_bot_username(self) -> str:
-        """Return this bot's live @username (lowercased, no leading ``@``).
-
-        Prefers the most recently observed handle over PTB's ``get_me()``
-        cache. ``Bot.username`` reads ``Bot._bot_user``, which is written only
-        by ``get_me()`` — after a BotFather rename it keeps returning the old
-        handle, so every mention comparison silently stops matching and the
-        exclusive-mention gate concludes the message is addressed to a
-        different bot. Observing the handle from inbound updates closes that
-        window without an extra Bot API round-trip.
-        """
-        observed = getattr(self, "_bot_username_observed", None)
-        if observed:
-            return observed
-        return (getattr(self._bot, "username", None) or "").lstrip("@").lower()
-
-    def _note_bot_username(self, username: Optional[str]) -> None:
-        """Record the bot's current @username, logging real renames."""
-        handle = (username or "").lstrip("@").lower()
-        if not handle:
-            return
-        previous = getattr(self, "_bot_username_observed", None)
-        if previous == handle:
-            return
-        self._bot_username_observed = handle
-        self._bot_identity_checked_at = time.monotonic()
-        if previous:
-            logger.info(
-                "[%s] Telegram bot username changed: @%s -> @%s "
-                "(mention routing now follows the new handle)",
-                self.name, previous, handle,
-            )
-
-    def _observe_bot_identity_from_message(self, message: Message) -> None:
-        """Learn our own handle from a message Telegram says we authored.
-
-        Telegram stamps the *current* username on the bot's own outgoing
-        messages and on ``reply_to_message`` when a user replies to us, so a
-        rename is observable from the update stream itself — no getMe needed.
-        Only trusted when the user id matches this bot, so another account's
-        handle can never be adopted as our own.
-        """
-        bot_id = getattr(self._bot, "id", None)
-        if bot_id is None:
-            return
-        for candidate in (
-            getattr(message, "from_user", None),
-            getattr(getattr(message, "reply_to_message", None), "from_user", None),
-        ):
-            if candidate is None:
-                continue
-            if getattr(candidate, "id", None) != bot_id:
-                continue
-            self._note_bot_username(getattr(candidate, "username", None))
-
-    def _bot_identity_is_fresh(self) -> bool:
-        """True when identity was re-read within the TTL.
-
-        ``None`` means never checked, which is always stale. Do not fold the
-        sentinel into ``0.0``: monotonic clocks have an arbitrary epoch that
-        can legitimately be smaller than the TTL on a freshly-booted host,
-        which would make "never" look like "just now".
-        """
-        checked_at = getattr(self, "_bot_identity_checked_at", None)
-        if checked_at is None:
-            return False
-        return (time.monotonic() - checked_at) < self._BOT_IDENTITY_TTL_SECONDS
-
-    async def _refresh_bot_identity(self, *, force: bool = False) -> None:
-        """Re-read the bot's identity from Telegram when the cache may be stale.
-
-        ``get_me()`` rewrites PTB's ``Bot._bot_user`` in place, so this also
-        repairs every other consumer of ``self._bot.username``. Best-effort:
-        a failed probe leaves the last known handle in place.
-        """
-        bot = self._bot
-        if bot is None or not callable(getattr(bot, "get_me", None)):
-            return
-        if not force and self._bot_identity_is_fresh():
-            return
-        try:
-            me = await asyncio.wait_for(bot.get_me(), self._BOT_IDENTITY_PROBE_TIMEOUT)
-        except asyncio.CancelledError:
-            raise
-        except Exception as exc:
-            logger.debug(
-                "[%s] Telegram identity refresh failed (keeping @%s): %s",
-                self.name, self._current_bot_username() or "unknown", exc,
-            )
-            return
-        self._bot_identity_checked_at = time.monotonic()
-        self._note_bot_username(getattr(me, "username", None))
-
     _BOT_IDENTITY_PROBE_TIMEOUT = 15.0
-
-    def _is_reply_to_bot(self, message: Message) -> bool:
-        if not self._bot or not getattr(message, "reply_to_message", None):
-            return False
-        reply_user = getattr(message.reply_to_message, "from_user", None)
-        return bool(reply_user and getattr(reply_user, "id", None) == getattr(self._bot, "id", None))
-
-    @classmethod
-    def _extract_bot_mention_usernames(cls, message: Message, self_username: str = "") -> set[str]:
-        """Extract explicit Telegram bot usernames mentioned in text/captions.
-
-        Foreign handles are only treated as bot mentions when they look
-        bot-shaped (``...bot``), which keeps human ``@handles`` from acting as
-        routing hints. ``self_username`` opts our OWN handle into the same set
-        regardless of shape: collectible (Fragment) usernames can be assigned
-        to bots and need not end in "bot" (@jarvis, @pic), and a bot addressed
-        by such a handle must still recognise itself.
-
-        Entity mentions are authoritative. The raw-text fallback is intentionally narrow so
-        entity-less mobile/client variants still work without treating email
-        addresses or arbitrary substrings as bot mentions.
-        """
-        mentioned_bot_usernames: set[str] = set()
-        own = (self_username or "").lstrip("@").lower()
-
-        def _is_bot_handle(handle: str) -> bool:
-            if not handle:
-                return False
-            if own and handle == own:
-                return True
-            return bool(cls._FOREIGN_BOT_HANDLE_RE.fullmatch(handle))
-
-        def _iter_sources():
-            yield getattr(message, "text", None) or "", getattr(message, "entities", None) or []
-            yield getattr(message, "caption", None) or "", getattr(message, "caption_entities", None) or []
-
-        for source_text, entities in _iter_sources():
-            for entity in entities:
-                entity_type = str(getattr(entity, "type", "")).split(".")[-1].lower()
-                if entity_type not in {"mention", "bot_command"}:
-                    continue
-                offset = int(getattr(entity, "offset", -1))
-                length = int(getattr(entity, "length", 0))
-                if offset < 0 or length <= 0:
-                    continue
-
-                entity_text = source_text[offset:offset + length].strip()
-                if entity_type == "mention":
-                    handle = entity_text.lstrip("@").lower()
-                    if _is_bot_handle(handle):
-                        mentioned_bot_usernames.add(handle)
-                    continue
-
-                # Telegram emits /cmd@botname as one bot_command entity, not as
-                # a separate mention entity. Treat that suffix as an explicit
-                # bot address for exclusive multi-bot routing even when the
-                # group has require_mention/free-response disabled.
-                at_index = entity_text.find("@")
-                if at_index < 0:
-                    continue
-                command_target = entity_text[at_index + 1:].strip().lower()
-                if _is_bot_handle(command_target):
-                    mentioned_bot_usernames.add(command_target)
-
-        # Entity-less fallback for older/client-specific updates. If Telegram
-        # supplied entities for a source, trust them and do not regex-rescue
-        # malformed/URL/code spans that the server did not mark as mentions.
-        for raw_text, entities in _iter_sources():
-            if not raw_text or entities:
-                continue
-            for match in re.finditer(r"(?i)(?<![A-Za-z0-9_`/])@([A-Za-z0-9_]{2,31})\b", raw_text):
-                handle = match.group(1).lower()
-                if _is_bot_handle(handle):
-                    mentioned_bot_usernames.add(handle)
-
-        return mentioned_bot_usernames
 
     def _message_mentions_bot(self, message: Message) -> bool:
         if not self._bot:
